Add base instructions for the spark-job-server image definition
This is an attempt to produce a Docker container that runs Spark Job
Server. It uses the official Java base container provided by Docker and
installs:

- Scala
- sbt
- Spark
- Spark Job Server

In a later refactor, creating a Spark base container and inheriting from
it may be a better approach.

Further instructions around getting things running can be found in the
`README`. There is also a mechanism to test the setup in the
accompanying `Makefile`.